requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
attrs==23.2.0
    # via
    #   -r requirements.in
    #   hypothesis
black==24.3.0
    # via
    #   -r requirements.in
    #   pytest-black
certifi==2024.2.2
    # via
    #   -r requirements.in
    #   requests
cffi==1.16.0
    # via
    #   -r requirements.in
    #   cryptography
charset-normalizer==3.3.2
    # via
    #   -r requirements.in
    #   requests
click==8.1.7
    # via
    #   -r requirements.in
    #   black
coreapi==2.3.3
    # via
    #   -r requirements.in
    #   django-rest-swagger
    #   openapi-codec
coreschema==0.0.4
    # via
    #   -r requirements.in
    #   coreapi
coverage[toml]==7.4.4
    # via
    #   -r requirements.in
    #   pytest-cov
cryptography==42.0.5
    # via
    #   -r requirements.in
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   -r requirements.in
    #   python3-openid
    #   social-auth-core
dj-database-url==2.1.0
    # via -r requirements.in
django==5.0.3
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   model-bakery
    #   social-auth-app-django
django-cors-headers==4.3.1
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filter==24.1
    # via -r requirements.in
django-rest-swagger==2.2.0
    # via -r requirements.in
django-templated-mail==1.1.1
    # via
    #   -r requirements.in
    #   djoser
djangorestframework==3.15.1
    # via
    #   -r requirements.in
    #   django-rest-swagger
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via
    #   -r requirements.in
    #   djoser
djoser==2.2.2
    # via -r requirements.in
drf-writable-nested==0.7.0
    # via -r requirements.in
execnet==2.0.2
    # via
    #   -r requirements.in
    #   pytest-xdist
faker==12.0.1
    # via
    #   -r requirements.in
    #   mixer
gunicorn==21.2.0
    # via -r requirements.in
hypothesis==6.99.13
    # via -r requirements.in
idna==3.6
    # via
    #   -r requirements.in
    #   requests
iniconfig==2.0.0
    # via
    #   -r requirements.in
    #   pytest
itypes==1.2.0
    # via
    #   -r requirements.in
    #   coreapi
jinja2==3.1.3
    # via
    #   -r requirements.in
    #   coreschema
markdown==3.6
    # via -r requirements.in
markdown-it-py==3.0.0
    # via
    #   -r requirements.in
    #   rich
markupsafe==2.1.5
    # via
    #   -r requirements.in
    #   jinja2
mdurl==0.1.2
    # via
    #   -r requirements.in
    #   markdown-it-py
mixer==7.2.2
    # via -r requirements.in
model-bakery==1.17.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   -r requirements.in
    #   black
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   social-auth-core
openapi-codec==1.3.2
    # via
    #   -r requirements.in
    #   django-rest-swagger
packaging==24.0
    # via
    #   -r requirements.in
    #   black
    #   gunicorn
    #   pytest
    #   pytest-sugar
pathspec==0.12.1
    # via
    #   -r requirements.in
    #   black
pillow==10.2.0
    # via -r requirements.in
platformdirs==4.2.0
    # via
    #   -r requirements.in
    #   black
pluggy==1.4.0
    # via
    #   -r requirements.in
    #   pytest
pprintpp==0.4.0
    # via
    #   -r requirements.in
    #   pytest-clarity
psycopg2==2.9.9
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pycparser==2.21
    # via
    #   -r requirements.in
    #   cffi
pygments==2.17.2
    # via
    #   -r requirements.in
    #   rich
pyjwt==2.8.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   social-auth-core
pytest==8.1.1
    # via
    #   -r requirements.in
    #   pytest-black
    #   pytest-clarity
    #   pytest-cov
    #   pytest-django
    #   pytest-sugar
    #   pytest-xdist
pytest-black==0.3.12
    # via -r requirements.in
pytest-clarity==1.0.1
    # via -r requirements.in
pytest-cov==5.0.0
    # via -r requirements.in
pytest-django==4.8.0
    # via -r requirements.in
pytest-sugar==1.0.0
    # via -r requirements.in
pytest-xdist==3.5.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   faker
python-dotenv==1.0.1
    # via -r requirements.in
python3-openid==3.2.0
    # via
    #   -r requirements.in
    #   social-auth-core
pytz==2024.1
    # via -r requirements.in
pyyaml==6.0.1
    # via -r requirements.in
requests==2.31.0
    # via
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via
    #   -r requirements.in
    #   social-auth-core
rest-framework-simplejwt==0.0.2
    # via -r requirements.in
rich==13.7.1
    # via
    #   -r requirements.in
    #   pytest-clarity
simplejson==3.19.2
    # via
    #   -r requirements.in
    #   django-rest-swagger
six==1.16.0
    # via
    #   -r requirements.in
    #   python-dateutil
social-auth-app-django==5.4.0
    # via
    #   -r requirements.in
    #   djoser
social-auth-core==4.5.3
    # via
    #   -r requirements.in
    #   social-auth-app-django
sortedcontainers==2.4.0
    # via
    #   -r requirements.in
    #   hypothesis
sqlparse==0.4.4
    # via
    #   -r requirements.in
    #   django
termcolor==2.4.0
    # via
    #   -r requirements.in
    #   pytest-sugar
toml==0.10.2
    # via
    #   -r requirements.in
    #   pytest-black
typing-extensions==4.10.0
    # via
    #   -r requirements.in
    #   dj-database-url
upgrade-requirements==1.7.0
    # via -r requirements.in
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   coreapi
urllib3==2.2.1
    # via
    #   -r requirements.in
    #   requests
whitenoise==6.6.0
    # via -r requirements.in
